pytorch_ver/requirements.txt
numpy
scikit-learn
scikit-image
umap-learn
matplotlib
pillow
tqdm